backend/services/chatbot_service/requirements.txt
torch
numpy
transformers
scikit-learn
requests
faiss-cpu
fastapi
pydantic
uvicorn
groq
supabase
sentence_transformers
